python/LbRelease/Nightlies directory removed (previous version of LHCb Nightlies system); fix for symbolic links to project directories on Linux
git-svn-id: http://svn.cern.ch/guest/lhcb/LbScripts/trunk@103274 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: db470d89c420a87aa8549705601ffc37ed89ad7a

+++ LbScripts/LbRelease/python/LbRelease/LHCbNightlies.py
@@ -181,7 +181,8 @@ class LHCbProjectBuilder(object):
         if not os.path.exists(self.generatePath(self.slot, self.project, 'TAG', self.project.getName())):
             os.makedirs(self.generatePath(self.slot, self.project, 'TAG', self.project.getName().upper()))
             if self.systemType != 'windows':
-                os.symlink(self.project.getName(), self.project.getName().upper())
+                if not os.path.exists(self.project.getName().upper()):
+                    os.symlink(self.project.getName(), self.project.getName().upper())
         os.chdir(self.generatePath(self.slot, self.project, 'TAG', self.project.getName()))
         waitcounter = 0
         while os.path.exists('checkout.working'):
